Move debugger attributes to shared partition (dotnet/coreclr#15100)
- Break down DebuggerAttributes.cs into more files, apply cleanup done in CoreRT
- Move two outliers (DebuggerStepperBoundaryAttribute, DebuggerVisualizerAttribute) from CoreFX to CoreLib so that all debugger attributes are together
