--- src/System.Private.CoreLib/src/System/Diagnostics/DebuggerAttributes.cs
@@ -1,361 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-/*============================================================
-**
-**
-**
-** Purpose: Attributes for debugger
-**
-**
-===========================================================*/
-
-using System;
-using System.Runtime.InteropServices;
-
-namespace System.Diagnostics
-{
-    //
-    // This attribute is used by the IL2IL toolchain to mark generated code to control debugger stepping policy
-    //
-    [System.Runtime.CompilerServices.DependencyReductionRoot]
-    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Method | AttributeTargets.Constructor, Inherited = false)]
-    public sealed class DebuggerStepThroughAttribute : Attribute
-    {
-        public DebuggerStepThroughAttribute() { }
-    }
-
-    [AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Method | AttributeTargets.Property | AttributeTargets.Constructor, Inherited = false)]
-    public sealed class DebuggerGuidedStepThroughAttribute : Attribute
-    {
-        public DebuggerGuidedStepThroughAttribute() { }
-    }
-
-    // This attribute is not in Win8P because it is not portable and so we aren't including it, if it turns out to be important we can consider 
-    // adding it into the next version of System.Diagnostics.Debug contract assembly
-    //[AttributeUsage(AttributeTargets.Method | AttributeTargets.Constructor, Inherited = false)]
-    //public sealed class DebuggerStepperBoundaryAttribute : Attribute
-    //{
-    //    public DebuggerStepperBoundaryAttribute() { }
-    //} 
-
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property | AttributeTargets.Constructor, Inherited = false)]
-    public sealed class DebuggerHiddenAttribute : Attribute
-    {
-        public DebuggerHiddenAttribute() { }
-    }
-
-    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method | AttributeTargets.Property | AttributeTargets.Constructor | AttributeTargets.Struct, Inherited = false)]
-    public sealed class DebuggerNonUserCodeAttribute : Attribute
-    {
-        public DebuggerNonUserCodeAttribute() { }
-    }
-
-    // Attribute class used by the compiler to mark modules.  
-    // If present, then debugging information for everything in the
-    // assembly was generated by the compiler, and will be preserved
-    // by the Runtime so that the debugger can provide full functionality
-    // in the case of JIT attach. If not present, then the compiler may
-    // or may not have included debugging information, and the Runtime
-    // won't preserve the debugging info, which will make debugging after
-    // a JIT attach difficult.
-    [AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Module, AllowMultiple = false)]
-    public sealed class DebuggableAttribute : Attribute
-    {
-        [Flags]
-        public enum DebuggingModes
-        {
-            None = 0x0,
-            Default = 0x1,
-            DisableOptimizations = 0x100,
-            IgnoreSymbolStoreSequencePoints = 0x2,
-            EnableEditAndContinue = 0x4
-        }
-
-        public DebuggableAttribute(bool isJITTrackingEnabled, bool isJITOptimizerDisabled)
-        {
-            DebuggingFlags = 0;
-
-            if (isJITTrackingEnabled)
-            {
-                DebuggingFlags |= DebuggingModes.Default;
-            }
-
-            if (isJITOptimizerDisabled)
-            {
-                DebuggingFlags |= DebuggingModes.DisableOptimizations;
-            }
-        }
-
-        public DebuggableAttribute(DebuggingModes modes)
-        {
-            DebuggingFlags = modes;
-        }
-
-        public bool IsJITTrackingEnabled => (DebuggingFlags & DebuggingModes.Default) != 0;
-
-        public bool IsJITOptimizerDisabled => (DebuggingFlags & DebuggingModes.DisableOptimizations) != 0;
-
-        public DebuggingModes DebuggingFlags { get; }
-    }
-
-    //  DebuggerBrowsableState states are defined as follows:
-    //      Never       never show this element
-    //      Expanded    expansion of the class is done, so that all visible internal members are shown
-    //      Collapsed   expansion of the class is not performed. Internal visible members are hidden
-    //      RootHidden  The target element itself should not be shown, but should instead be 
-    //                  automatically expanded to have its members displayed.
-    //  Default value is collapsed
-
-    //  Please also change the code which validates DebuggerBrowsableState variable (in this file)
-    //  if you change this enum.
-    public enum DebuggerBrowsableState
-    {
-        Never = 0,
-        //Expanded is not supported in this release
-        //Expanded = 1, 
-        Collapsed = 2,
-        RootHidden = 3
-    }
-
-
-    // the one currently supported with the csee.dat 
-    // (mcee.dat, autoexp.dat) file. 
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property, AllowMultiple = false)]
-    public sealed class DebuggerBrowsableAttribute : Attribute
-    {
-        private DebuggerBrowsableState _state;
-        public DebuggerBrowsableAttribute(DebuggerBrowsableState state)
-        {
-            if (state < DebuggerBrowsableState.Never || state > DebuggerBrowsableState.RootHidden)
-                throw new ArgumentOutOfRangeException(nameof(state));
-
-            _state = state;
-        }
-        public DebuggerBrowsableState State
-        {
-            get { return _state; }
-        }
-    }
-
-
-    // DebuggerTypeProxyAttribute
-    [AttributeUsage(AttributeTargets.Struct | AttributeTargets.Class | AttributeTargets.Assembly, AllowMultiple = true)]
-    public sealed class DebuggerTypeProxyAttribute : Attribute
-    {
-        private string _typeName;
-        private string _targetName;
-        private Type _target;
-
-        public DebuggerTypeProxyAttribute(Type type)
-        {
-            if (type == null)
-            {
-                throw new ArgumentNullException(nameof(type));
-            }
-
-            _typeName = type.AssemblyQualifiedName;
-        }
-
-        public DebuggerTypeProxyAttribute(string typeName)
-        {
-            _typeName = typeName;
-        }
-        public string ProxyTypeName
-        {
-            get { return _typeName; }
-        }
-
-        public Type Target
-        {
-            set
-            {
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                _targetName = value.AssemblyQualifiedName;
-                _target = value;
-            }
-
-            get { return _target; }
-        }
-
-        public string TargetTypeName
-        {
-            get { return _targetName; }
-            set { _targetName = value; }
-        }
-    }
-
-    // This attribute is used to control what is displayed for the given class or field 
-    // in the data windows in the debugger.  The single argument to this attribute is
-    // the string that will be displayed in the value column for instances of the type.  
-    // This string can include text between { and } which can be either a field, 
-    // property or method (as will be documented in mscorlib).  In the C# case, 
-    // a general expression will be allowed which only has implicit access to the this pointer 
-    // for the current instance of the target type. The expression will be limited, 
-    // however: there is no access to aliases, locals, or pointers. 
-    // In addition, attributes on properties referenced in the expression are not processed.
-    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Delegate | AttributeTargets.Enum | AttributeTargets.Field | AttributeTargets.Property | AttributeTargets.Assembly, AllowMultiple = true)]
-    public sealed class DebuggerDisplayAttribute : Attribute
-    {
-        private string _name;
-        private string _value;
-        private string _type;
-        private string _targetName;
-        private Type _target;
-
-        public DebuggerDisplayAttribute(string value)
-        {
-            if (value == null)
-            {
-                _value = "";
-            }
-            else
-            {
-                _value = value;
-            }
-            _name = "";
-            _type = "";
-        }
-
-        public string Value
-        {
-            get { return _value; }
-        }
-
-        public string Name
-        {
-            get { return _name; }
-            set { _name = value; }
-        }
-
-        public string Type
-        {
-            get { return _type; }
-            set { _type = value; }
-        }
-
-        public Type Target
-        {
-            set
-            {
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                _targetName = value.AssemblyQualifiedName;
-                _target = value;
-            }
-            get { return _target; }
-        }
-
-        public string TargetTypeName
-        {
-            get { return _targetName; }
-            set { _targetName = value; }
-        }
-    }
-    ///// <summary>
-    ///// Signifies that the attributed type has a visualizer which is pointed
-    ///// to by the parameter type name strings.
-    ///// </summary>
-    //[AttributeUsage(AttributeTargets.Struct | AttributeTargets.Class | AttributeTargets.Assembly, AllowMultiple = true)]
-    //public sealed class DebuggerVisualizerAttribute: Attribute
-    //{
-    //    private string visualizerObjectSourceName;
-    //    private string visualizerName;
-    //    private string description;
-    //    private string targetName;
-    //    private Type target;
-
-    //    public DebuggerVisualizerAttribute(string visualizerTypeName)
-    //    {
-    //        this.visualizerName = visualizerTypeName;
-    //    }
-    //    public DebuggerVisualizerAttribute(string visualizerTypeName, string visualizerObjectSourceTypeName)
-    //    {
-    //        this.visualizerName = visualizerTypeName;
-    //        this.visualizerObjectSourceName = visualizerObjectSourceTypeName;
-    //    }
-    //    public DebuggerVisualizerAttribute(string visualizerTypeName, Type visualizerObjectSource)
-    //    {
-    //        if (visualizerObjectSource == null) {
-    //            throw new ArgumentNullException("visualizerObjectSource");
-    //        }
-    //        Contract.EndContractBlock();
-    //        this.visualizerName = visualizerTypeName;
-    //        this.visualizerObjectSourceName = visualizerObjectSource.AssemblyQualifiedName;
-    //    }
-    //    public DebuggerVisualizerAttribute(Type visualizer)
-    //    {    
-    //        if (visualizer == null) {
-    //            throw new ArgumentNullException("visualizer");
-    //        }
-    //        Contract.EndContractBlock();
-    //        this.visualizerName = visualizer.AssemblyQualifiedName;
-    //    }
-    //    public DebuggerVisualizerAttribute(Type visualizer, Type visualizerObjectSource)
-    //    {
-    //        if (visualizer == null) {
-    //            throw new ArgumentNullException("visualizer");
-    //        }
-    //        if (visualizerObjectSource == null) {
-    //            throw new ArgumentNullException("visualizerObjectSource");
-    //        }
-    //        Contract.EndContractBlock();
-    //        this.visualizerName = visualizer.AssemblyQualifiedName;
-    //        this.visualizerObjectSourceName = visualizerObjectSource.AssemblyQualifiedName;
-    //    }
-    //    public DebuggerVisualizerAttribute(Type visualizer, string visualizerObjectSourceTypeName)
-    //    {
-    //        if (visualizer == null) {
-    //            throw new ArgumentNullException("visualizer");
-    //        }
-    //        Contract.EndContractBlock();
-    //        this.visualizerName = visualizer.AssemblyQualifiedName;
-    //        this.visualizerObjectSourceName = visualizerObjectSourceTypeName;
-    //    }
-
-    //    public string VisualizerObjectSourceTypeName
-    //    {
-    //        get { return visualizerObjectSourceName; }
-    //    }
-    //    public string VisualizerTypeName
-    //    {
-    //        get { return visualizerName; }
-    //    }
-    //    public string Description
-    //    {
-    //        get { return description; }
-    //        set { description = value; }
-    //    }
-
-    //    public Type Target
-    //    {
-    //        set {                 
-    //            if( value == null) {
-    //                throw new ArgumentNullException("value");
-    //            }
-    //            Contract.EndContractBlock();
-
-    //            targetName = value.AssemblyQualifiedName; 
-    //            target = value; 
-    //        }
-
-    //        get { return target; }
-    //    }
-
-    //    public string TargetTypeName
-    //    {
-    //        set { targetName = value; }
-    //        get { return targetName; }
-    //    }
-    //}
-}
-
-
